Add harden_linux_required_packages_user
Adding harden_linux_required_packages_user to enable users of this role
only to specify the packages they require in addition to the
defaults specified in the role. The two lists are merged together and
the combined package list is installed.

If no harden_linux_required_packages_user is defined, only the default
list is installed.